added a bit more of the ufuncs, started BZ
