+++ src/sisl/_core/geometry.py
@@ -774,32 +774,6 @@ class Geometry(
             with get_sile(sile, mode="r") as fh:
                 return fh.read_geometry(*args, **kwargs)
 
-    def write(self, sile: SileLike, *args, **kwargs) -> None:
-        """Writes geometry to the `Sile` using `sile.write_geometry`
-
-        Parameters
-        ----------
-        sile :
-            a `Sile` object which will be used to write the geometry
-            if it is a string it will create a new sile using `get_sile`
-        *args, **kwargs:
-            Any other args will be passed directly to the
-            underlying routine
-
-        See Also
-        --------
-        read : reads a `Geometry` from a given `Sile`/file
-        """
-        # This only works because, they *must*
-        # have been imported previously
-        from sisl.io import BaseSile, get_sile
-
-        if isinstance(sile, BaseSile):
-            sile.write_geometry(self, *args, **kwargs)
-        else:
-            with get_sile(sile, mode="w") as fh:
-                fh.write_geometry(self, *args, **kwargs)
-
     def __str__(self) -> str:
         """str of the object"""
         s = f"{self.__class__.__name__}{{na: {self.na}, no: {self.no},\n "
@@ -1301,522 +1275,6 @@ class Geometry(
             other_extend(idx)
         return _a.arrayi(idx_self), _a.arrayi(idx_other)
 
-    def sort(self, **kwargs) -> Union[Geometry, Tuple[Geometry, List]]:
-        r"""Sort atoms in a nested fashion according to various criteria
-
-        There are many ways to sort a `Geometry`.
-        - by Cartesian coordinates, `axis`
-        - by lattice vectors, `lattice`
-        - by user defined vectors, `vector`
-        - by grouping atoms, `group`
-        - by a user defined function, `func`
-        - by a user defined function using internal sorting algorithm, `func_sort`
-
-        - a combination of the above in arbitrary order
-
-        Additionally one may sort ascending or descending.
-
-        This method allows nested sorting based on keyword arguments.
-
-        Parameters
-        ----------
-        atoms : int or array_like, optional
-           only perform sorting algorithm for subset of atoms. This is *NOT* a positional dependent
-           argument. All sorting algorithms will _only_ be performed on these atoms.
-           Default, all atoms will be sorted.
-        ret_atoms : bool, optional
-           return a list of list for the groups of atoms that have been sorted.
-        axis : int or tuple of int, optional
-           sort coordinates according to Cartesian coordinates, if a tuple of
-           ints is passed it will be equivalent to ``sort(axis0=axis[0], axis1=axis[1])``.
-           This behaves differently than `numpy.lexsort`!
-        lattice : int or tuple of int, optional
-           sort coordinates according to lattice vectors, if a tuple of
-           ints is passed it will be equivalent to ``sort(lattice0=lattice[0], lattice1=lattice[1])``.
-           Note that before sorting we multiply the fractional coordinates by the length of the
-           lattice vector. This ensures that `atol` is meaningful for both `axis` and `lattice` since
-           they will be on the same order of magnitude.
-           This behaves differently than `numpy.lexsort`!
-        vector : (3, ), optional
-           sort along a user defined vector, similar to `lattice` but with a user defined
-           direction. Note that `lattice` sorting and `vector` sorting are *only* equivalent
-           when the lattice vector is orthogonal to the other lattice vectors.
-        group : {'Z', 'symbol', 'tag', 'species'} or (str, ...), optional
-           group together a set of atoms by various means.
-           `group` may be one of the listed strings.
-           For ``'Z'`` atoms will be grouped in atomic number
-           For ``'symbol'`` atoms will be grouped by their atomic symbol.
-           For ``'tag'`` atoms will be grouped by their atomic tag.
-           For ``'species'`` atoms will be sorted according to their specie index.
-           If a tuple/list is passed the first item is described. All subsequent items are a
-           list of groups, where each group comprises elements that should be sorted on an
-           equal footing. If one of the groups is None, that group will be replaced with all
-           non-mentioned elements. See examples.
-        func : callable or list-like of callable, optional
-           pass a sorting function which should have an interface like ``func(geometry, atoms, **kwargs)``.
-           The first argument is the geometry to sort. The 2nd argument is a list of indices in
-           the current group of sorted atoms. And ``**kwargs`` are any optional arguments
-           currently collected, i.e. `ascend`, `atol` etc.
-           The function should return either a list of atoms, or a list of list of atoms (in which
-           case the atomic indices have been split into several groups that will be sorted individually
-           for subsequent sorting methods).
-           In either case the returned indices must never hold any other indices but the ones passed
-           as ``atoms``.
-           If a list/tuple of functions, they will be processed in that order.
-        func_sort : callable or list-like of callable, optional
-           pass a function returning a 1D array corresponding to all atoms in the geometry.
-           The interface should simply be: ``func(geometry)``.
-           Those values will be passed down to the internal sorting algorithm.
-           To be compatible with `atol` the returned values from `func_sort` should
-           be on the scale of coordinates (in Ang).
-        ascend, descend : bool, optional
-            control ascending or descending sorting for all subsequent sorting methods.
-            Default ``ascend=True``.
-        atol : float, optional
-            absolute tolerance when sorting numerical arrays for subsequent sorting methods.
-            When a selection of sorted coordinates are grouped via `atol`, we ensure such
-            a group does not alter its indices. I.e. the group is *always* ascending indices.
-            Note this may have unwanted side-effects if `atol` is very large compared
-            to the difference between atomic coordinates.
-            Default ``1e-9``.
-
-        Notes
-        -----
-        The order of arguments is also the sorting order. ``sort(axis=0, lattice=0)`` is different
-        from ``sort(lattice=0, axis=0)``
-
-        All arguments may be suffixed with integers. This allows multiple keyword arguments
-        to control sorting algorithms
-        in different order. It also allows changing of sorting settings between different calls.
-        Note that the integers have no relevance to the order of execution!
-        See examples.
-
-        Returns
-        -------
-        geometry : Geometry
-            sorted geometry
-        index : list of list of indices
-            indices that would sort the original structure to the output, only returned if `ret_atoms` is True
-
-        Examples
-        --------
-        >>> geom = sisl.geom.bilayer(top_atoms=sisl.Atom[5, 7], bottom_atoms=sisl.Atom(6))
-        >>> geom = geom.tile(5, 0).repeat(7, 1)
-
-        Sort according to :math:`x` coordinate
-
-        >>> geom.sort(axis=0)
-
-        Sort according to :math:`z`, then :math:`x` for each group created from first sort
-
-        >>> geom.sort(axis=(2, 0))
-
-        Sort according to :math:`z`, then first lattice vector
-
-        >>> geom.sort(axis=2, lattice=0)
-
-        Sort according to :math:`z` (ascending), then first lattice vector (descending)
-
-        >>> geom.sort(axis=2, ascend=False, lattice=0)
-
-        Sort according to :math:`z` (descending), then first lattice vector (ascending)
-        Note how integer suffixes has no importance.
-
-        >>> geom.sort(ascend1=False, axis=2, ascend0=True, lattice=0)
-
-        Sort only atoms ``range(1, 5)`` first by :math:`z`, then by first lattice vector
-
-        >>> geom.sort(axis=2, lattice=0, atoms=np.arange(1, 5))
-
-        Sort two groups of atoms ``[range(1, 5), range(5, 10)]`` (individually) by :math:`z` coordinate
-
-        >>> geom.sort(axis=2, atoms=[np.arange(1, 5), np.arange(5, 10)])
-
-        The returned sorting indices may be used for manual sorting. Note
-        however, that this requires one to perform a sorting for all atoms.
-        In such a case the following sortings are equal.
-
-        >>> geom0, atoms0 = geom.sort(axis=2, lattice=0, ret_atoms=True)
-        >>> _, atoms1 = geom.sort(axis=2, ret_atoms=True)
-        >>> geom1, atoms1 = geom.sort(lattice=0, atoms=atoms1, ret_atoms=True)
-        >>> geom2 = geom.sub(np.concatenate(atoms0))
-        >>> geom3 = geom.sub(np.concatenate(atoms1))
-        >>> assert geom0 == geom1
-        >>> assert geom0 == geom2
-        >>> assert geom0 == geom3
-
-        Default sorting is equivalent to ``axis=(0, 1, 2)``
-
-        >>> assert geom.sort() == geom.sort(axis=(0, 1, 2))
-
-        Sort along a user defined vector ``[2.2, 1., 0.]``
-
-        >>> geom.sort(vector=[2.2, 1., 0.])
-
-        Integer specification has no influence on the order of operations.
-        It is _always_ the keyword argument order that determines the operation.
-
-        >>> assert geom.sort(axis2=1, axis0=0, axis1=2) == geom.sort(axis=(1, 0, 2))
-
-        Sort by atomic numbers
-
-        >>> geom.sort(group='Z') # 5, 6, 7
-
-        One may group several elements together on an equal footing (``None`` means all non-mentioned elements)
-        The order of the groups are important (the first two are _not_ equal, the last three _are_ equal)
-
-        >>> geom.sort(group=('symbol', 'C'), axis=2) # C will be sorted along z
-        >>> geom.sort(axis=1, atoms='C', axis1=2) # all along y, then C sorted along z
-        >>> geom.sort(group=('symbol', 'C', None)) # C, [B, N]
-        >>> geom.sort(group=('symbol', None, 'C')) # [B, N], C
-        >>> geom.sort(group=('symbol', ['N', 'B'], 'C')) # [B, N], C (B and N unaltered order)
-        >>> geom.sort(group=('symbol', ['B', 'N'], 'C')) # [B, N], C (B and N unaltered order)
-
-        A group based sorting can use *anything* that can be fetched from the `Atom` object,
-        sort first according to mass, then for all with the same mass, sort according to atomic
-        tag:
-
-        >>> geom.sort(group0='mass', group1='tag')
-
-        A too high `atol` may have unexpected side-effects. This is because of the way
-        the sorting algorithm splits the sections for nested sorting.
-        So for coordinates with a continuous displacement the sorting may break and group
-        a large range into 1 group. Consider the following array to be split in groups
-        while sorting.
-
-        An example would be a linear chain with a middle point with a much shorter distance.
-
-        >>> x = np.arange(5) * 0.1
-        >>> x[3:] -= 0.095
-        y = z = np.zeros(5)
-        geom = si.Geometry(np.stack((x, y, z), axis=1))
-        >>> geom.xyz[:, 0]
-        [0.    0.1   0.2   0.205 0.305]
-
-        In this case a high tolerance (``atol>0.005``) would group atoms 2 and 3
-        together
-
-        >>> geom.sort(atol=0.01, axis=0, ret_atoms=True)[1]
-        [[0], [1], [2, 3], [4]]
-
-        However, a very low tolerance will not find these two as atoms close
-        to each other.
-
-        >>> geom.sort(atol=0.001, axis=0, ret_atoms=True)[1]
-        [[0], [1], [2], [3], [4]]
-        """
-
-        # We need a way to easily handle nested lists
-        # This small class handles lists and allows appending nested lists
-        # while flattening them.
-        class NestedList:
-            __slots__ = ("_idx",)
-
-            def __init__(self, idx=None, sort=False):
-                self._idx = []
-                if not idx is None:
-                    self.append(idx, sort)
-
-            def append(self, idx, sort=False):
-                if isinstance(idx, (tuple, list, ndarray)):
-                    if isinstance(idx[0], (tuple, list, ndarray)):
-                        for ix in idx:
-                            self.append(ix, sort)
-                        return
-                elif isinstance(idx, NestedList):
-                    idx = idx.tolist()
-                if len(idx) > 0:
-                    if sort:
-                        self._idx.append(np.sort(idx))
-                    else:
-                        self._idx.append(np.asarray(idx))
-
-            def __iter__(self):
-                yield from self._idx
-
-            def __len__(self):
-                return len(self._idx)
-
-            def ravel(self):
-                if len(self) == 0:
-                    return np.array([], dtype=np.int64)
-                return concatenate([i for i in self]).ravel()
-
-            def tolist(self):
-                return self._idx
-
-            def __str__(self):
-                if len(self) == 0:
-                    return f"{self.__class__.__name__}{{empty}}"
-                out = ",\n ".join(map(lambda x: str(x.tolist()), self))
-                return f"{self.__class__.__name__}{{\n {out}}}"
-
-        def _sort(val, atoms, **kwargs):
-            """We do not sort according to lexsort"""
-            if len(val) <= 1:
-                # no values to sort
-                return atoms
-
-            # control ascend vs descending
-            ascend = kwargs["ascend"]
-            atol = kwargs["atol"]
-
-            new_atoms = NestedList()
-            for atom in atoms:
-                if len(atom) <= 1:
-                    # no need for complexity
-                    new_atoms.append(atom)
-                    continue
-
-                # Sort values
-                jdx = atom[argsort(val[atom])]
-                if ascend:
-                    d = diff(val[jdx]) > atol
-                else:
-                    jdx = jdx[::-1]
-                    d = diff(val[jdx]) < -atol
-                new_atoms.append(split(jdx, d.nonzero()[0] + 1), sort=True)
-            return new_atoms
-
-        # Functions allowed by external users
-        funcs = dict()
-
-        def _axis(axis, atoms, **kwargs):
-            """Cartesian coordinate sort"""
-            if isinstance(axis, int):
-                axis = (axis,)
-            for ax in axis:
-                atoms = _sort(self.xyz[:, ax], atoms, **kwargs)
-            return atoms
-
-        funcs["axis"] = _axis
-
-        def _lattice(lattice, atoms, **kwargs):
-            """
-            We scale the fractional coordinates with the lattice vector length.
-            This ensures `atol` has a meaningful size for very large structures.
-            """
-            if isinstance(lattice, int):
-                lattice = (lattice,)
-            fxyz = self.fxyz
-            for ax in lattice:
-                atoms = _sort(fxyz[:, ax] * self.lattice.length[ax], atoms, **kwargs)
-            return atoms
-
-        funcs["lattice"] = _lattice
-
-        def _vector(vector, atoms, **kwargs):
-            """
-            Calculate fraction of positions along a vector and sort along it.
-            We first normalize the vector to ensure that `atol` is meaningful
-            for very large structures (ensures scale is on the order of Ang).
-
-            A vector projection will only be equivalent to lattice projection
-            when a lattice vector is orthogonal to the other lattice vectors.
-            """
-            # Ensure we are using a copied data array
-            vector = _a.asarrayd(vector).copy()
-            # normalize
-            vector /= fnorm(vector)
-            # Perform a . b^ == scalar projection
-            return _sort(self.xyz.dot(vector), atoms, **kwargs)
-
-        funcs["vector"] = _vector
-
-        def _funcs(funcs, atoms, **kwargs):
-            """
-            User defined function (tuple/list of function)
-            """
-
-            def _func(func, atoms, kwargs):
-                nl = NestedList()
-                for atom in atoms:
-                    # TODO add check that
-                    #  res = func(...) in a
-                    # A user *may* remove an atom from the sorting here (but
-                    # that negates all sorting of that atom)
-                    nl.append(func(self, atom, **kwargs))
-                return nl
-
-            if callable(funcs):
-                funcs = [funcs]
-            for func in funcs:
-                atoms = _func(func, atoms, kwargs)
-            return atoms
-
-        funcs["func"] = _funcs
-
-        def _func_sort(funcs, atoms, **kwargs):
-            """
-            User defined function, but using internal sorting
-            """
-            if callable(funcs):
-                funcs = [funcs]
-            for func in funcs:
-                atoms = _sort(func(self), atoms, **kwargs)
-            return atoms
-
-        funcs["func_sort"] = _func_sort
-
-        def _group_vals(vals, groups, atoms, **kwargs):
-            """
-            vals should be of size len(self) and be parsable
-            by numpy
-            """
-            nl = NestedList()
-
-            # Create unique list of values
-            uniq_vals = np.unique(vals)
-            if len(groups) == 0:
-                # fake the groups argument
-                groups = [[i] for i in uniq_vals]
-            else:
-                # Check if one of the elements of group is None
-                # In this case we replace it with the missing rest
-                # of the missing unique items
-                try:
-                    none_idx = groups.index(None)
-
-                    # we have a None (ensure we use a list, tuples are
-                    # immutable)
-                    groups = list(groups)
-                    groups[none_idx] = []
-
-                    uniq_groups = np.unique(concatenate(groups))
-                    # add a new group that is in uniq_vals, but not in uniq_groups
-                    rest = uniq_vals[isin(uniq_vals, uniq_groups, invert=True)]
-                    groups[none_idx] = rest
-                except ValueError:
-                    # there is no None in the list
-                    pass
-
-            for at in atoms:
-                # reduce search
-                at_vals = vals[at]
-                # loop group values
-                for group in groups:
-                    # retain original indexing
-                    nl.append(at[isin(at_vals, group)])
-            return nl
-
-        def _group(method_group, atoms, **kwargs):
-            """
-            Group based sorting is based on a named identification.
-
-            group: str or tuple of (str, list of lists)
-
-            symbol: order by symbol (most cases same as Z)
-            Z: order by atomic number
-            tag: order by atom tag (should be the same as specie)
-            specie/species: order by specie (in order of contained in the Geometry)
-            """
-            # Create new list
-            nl = NestedList()
-
-            if isinstance(method_group, str):
-                method = method_group
-                groups = []
-            elif isinstance(method_group[0], str):
-                method, *in_groups = method_group
-
-                # Ensure all groups are lists
-                groups = []
-                NoneType = type(None)
-                for group in in_groups:
-                    if isinstance(group, (tuple, list, ndarray, NoneType)):
-                        groups.append(group)
-                    else:
-                        groups.append([group])
-            else:
-                # a special case where group is a list of lists
-                # i.e. [[0, 1, 2], [3, 4, 5]]
-                for idx in method_group:
-                    idx = self._sanitize_atoms(idx)
-                    for at in atoms:
-                        nl.append(at[isin(at, idx)])
-                return nl
-
-            # See if the attribute exists for the atoms
-            if method.lower() == "species":
-                # this one has two spelling options!
-                method = "specie"
-
-            # now get them through `getattr`
-            if hasattr(self.atoms[0], method):
-                vals = [getattr(a, method) for a in self.atoms]
-
-            elif hasattr(self.atoms[0], method.lower()):
-                method = method.lower()
-                vals = [getattr(a, method) for a in self.atoms]
-
-            else:
-                raise ValueError(
-                    f"{self.__class__.__name__}.sort group only supports attributes that can be fetched from Atom objects, some are [Z, species, tag, symbol, mass, ...] and more"
-                )
-
-            return _group_vals(np.array(vals), groups, atoms, **kwargs)
-
-        funcs["group"] = _group
-
-        def stripint(s):
-            """Remove integers from end of string -> Allow multiple arguments"""
-            if s[-1] in "0123456789":
-                return stripint(s[:-1])
-            return s
-
-        # Now perform cumultative sort function
-        # Our point is that we would like to allow users to do consecutive sorting
-        # based on different keys
-
-        # We also allow specific keys for specific methods
-        func_kw = dict()
-        func_kw["ascend"] = True
-        func_kw["atol"] = 1e-9
-
-        def update_flag(kw, arg, val):
-            if arg in ["ascending", "ascend"]:
-                kw["ascend"] = val
-                return True
-            elif arg in ["descending", "descend"]:
-                kw["ascend"] = not val
-                return True
-            elif arg == "atol":
-                kw["atol"] = val
-                return True
-            return False
-
-        # Default to all atoms
-        atoms = NestedList(self._sanitize_atoms(kwargs.pop("atoms", None)))
-        ret_atoms = kwargs.pop("ret_atoms", False)
-
-        # In case the user just did geometry.sort, it will default to sort x, y, z
-        if len(kwargs) == 0:
-            kwargs["axis"] = (0, 1, 2)
-
-        for key_int, method in kwargs.items():
-            key = stripint(key_int)
-            if update_flag(func_kw, key, method):
-                continue
-            if not key in funcs:
-                raise ValueError(
-                    f"{self.__class__.__name__}.sort unrecognized keyword '{key}' ('{key_int}')"
-                )
-            # call sorting algorithm and retrieve new grouped sorting
-            atoms = funcs[key](method, atoms, **func_kw)
-
-        # convert to direct list
-        atoms_flat = atoms.ravel()
-
-        # Ensure that all atoms are present
-        if len(atoms_flat) != len(self):
-            all_atoms = _a.arangei(len(self))
-            all_atoms[np.sort(atoms_flat)] = atoms_flat[:]
-            atoms_flat = all_atoms
-
-        if ret_atoms:
-            return self.sub(atoms_flat), atoms.tolist()
-        return self.sub(atoms_flat)
-
     def optimize_nsc(
         self,
         axes: Optional[Union[int, Sequence[int]]] = None,
@@ -2284,27 +1742,6 @@ class Geometry(
             g.xyz[idx] = fxyz[idx] @ self.cell
         return g
 
-    def swap(self, atoms_a: AtomsArgument, atoms_b: AtomsArgument) -> Geometry:
-        """Swap a set of atoms in the geometry and return a new one
-
-        This can be used to reorder elements of a geometry.
-
-        Parameters
-        ----------
-        atoms_a : array_like
-             the first list of atomic coordinates
-        atoms_b : array_like
-             the second list of atomic coordinates
-        """
-        atoms_a = self._sanitize_atoms(atoms_a)
-        atoms_b = self._sanitize_atoms(atoms_b)
-        xyz = np.copy(self.xyz)
-        xyz[atoms_a, :] = self.xyz[atoms_b, :]
-        xyz[atoms_b, :] = self.xyz[atoms_a, :]
-        return self.__class__(
-            xyz, atoms=self.atoms.swap(atoms_a, atoms_b), lattice=self.lattice.copy()
-        )
-
     def add_vacuum(
         self, vacuum: float, axis: int, offset: Sequence[float] = (0, 0, 0)
     ) -> Geometry:
@@ -2336,36 +1773,6 @@ class Geometry(
             nsc[axis] = 1
             new.lattice.set_nsc(nsc)
         return new
-
-    def insert(self, atom: int, other: GeometryLike) -> Geometry:
-        """Inserts other atoms right before index
-
-        We insert the `geometry` `Geometry` before `atom`.
-        Note that this will not change the unit cell.
-
-        Parameters
-        ----------
-        atom :
-           the atomic index at which the other geometry is inserted
-        other :
-           the other geometry to be inserted
-
-        See Also
-        --------
-        add : add geometries
-        append : appending geometries
-        prepend : prending geometries
-        attach : attach a geometry
-        """
-        atom = self._sanitize_atoms(atom)
-        if atom.size > 1:
-            raise ValueError(
-                f"{self.__class__.__name__}.insert requires only 1 atomic index for insertion."
-            )
-        other = self.new(other)
-        xyz = np.insert(self.xyz, atom, other.xyz, axis=0)
-        atoms = self.atoms.insert(atom, other.atoms)
-        return self.__class__(xyz, atoms, lattice=self.lattice.copy())
 
     def __add__(self, b) -> Geometry:
         """Merge two geometries (or geometry and supercell)
